+++ helpers/tilemaphelper.hpp
@@ -10,7 +10,7 @@ class TileMapHelper {
 public:
     static TileMap create_tilemap();
     static void add_room_to(TileMap& tilemap, const Room& room);
-    static int rooms_count_of(const TileMap& tilemap);
+    static size_t rooms_count_of(const TileMap& tilemap);
     static int narrow_rooms_of(const TileMap& tilemap);
     static int tiny_rooms_of(const TileMap& tilemap);
     static string to_painted_map_string(const TileMap& tilemap);
